feat: added laptop model for intro
